requirements.txt
python-dotenv
openai==1.2.0
pyaudio==0.2.14
instructor==0.3.1
textual==0.41.0
pygame